refactor(client/src/component/Profile - /HomeAdmin): reacomodado en Profile - aplicado de estilos HomeAdmin
