+++ client/src/App.js
@@ -7,12 +7,11 @@ import NavBar from "./components/NavBar/NavBar";
 import { Users } from "./components/Users/Users.jsx";
 import { FormScore } from "./components/FormScore/FormScore.jsx";
 import { FormSponsor } from "./components/FormSponsor/FormSponsor.jsx";
-import { FormUser } from "./components/FormUser/FormUser.jsx";
 import { FormTournament } from "./components/Tournament/FormTournament.jsx";
 import  LogIn  from "./components/LogIn/LogIn.jsx";
-import {FormUserRegister} from "./components/FormUser copy/FormUserRegister.jsx"
-import ControlPanel from "./controlPanel/ControlPanel.jsx";
-import ProfileUser from "./components/ProfileUser/ProfileUser.jsx";
+import {FormUserRegister} from "./components/FormUserRegister/FormUserRegister.jsx"
+import ControlPanel from "./components/controlPanel/ControlPanel.jsx";
+import ProfileUser from "./components/Profile/ProfileUser/ProfileUser.jsx";
 
 
 function App() {
@@ -25,7 +24,6 @@ function App() {
       <Route exact path={"/Users/:userId"} component={DetailsUser} />
       <Route exact path={"/CreateScore/:userId"} component={FormScore} />
       <Route exact path={"/sponsor"} component={FormSponsor}/>
-      <Route exact path={"/EditProfile/:userId"} component={FormUser}/>
       <Route exact path={"/tournament"} component={FormTournament} />
       <Route exact path={"/login"} component={LogIn} />
       <Route exact path={"/cpanel"} component={ControlPanel}/>
